requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
beepy==1.0.7
bleach==6.1.0
blinker==1.8.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
cloudscraper==1.2.71
comm==0.2.1
copulas==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
finnhub-python==2.4.19
flake8==7.0.0
Flask==3.0.3
flatbuffers==25.2.10
frozendict==2.4.6
gast==0.6.0
gevent==23.9.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
idna==3.6
importlib_metadata==8.4.0
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keras==3.8.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
msgpack==1.0.2
multidict==6.0.5
multitasking==0.0.11
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.9.2
nest-asyncio==1.5.8
numpy==1.26.3
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pandas_ta==0.3.14b0
pandocfilters==1.5.1
parso==0.8.3
patsy==1.0.1
peewee==3.17.8
pexpect==4.9.0
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.4.0
ppi_client==1.2.1
prompt-toolkit==3.0.43
protobuf==5.29.3
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyCocos==0.2.12
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.17.2
pyotp==2.9.0
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.16.2
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.12.0
selenium==4.29.0
signalrcorePPI==0.9.2
simpleaudio==1.0.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2023.4
undetected-chromedriver==3.5.5
urllib3==2.1.0
vcrpy==6.0.1
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.51
zipp==3.20.0
zope.event==5.0
zope.interface==6.1
